third_party/WebKit/LayoutTests/http/tests/loading/sxg/sxg-non-secure-origin.html

<!DOCTYPE html>
<title>Loading SignedHTTPExchange from non secure origin must fail</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  await waitUntilDidFinishLoadForFrame;

  // The timestamp of the test SXG file is "Apr 1 2018 00:00 UTC" and valid
  // until "Apr 8 2018 00:00 UTC".
  await setSignedExchangeVerificationTime(new Date("Apr 1 2018 00:01 UTC"));

  try {
    await openSXGInIframeAndWaitForMessage(
        t, 'http://example.test:8080/loading/sxg/resources/sxg-location.sxg');
    assert_unreached('SXG should not have loaded');
  } catch (e) {
    assert_equals(e, 'timeout');
  }
}, 'Loading SignedHTTPExchange from non secure origin must fail');

</script>
</body>
